Apps Icon Opacity + SearchBox Modification
Changed the App Icon opacity to 0.5 to make it look more like the Google one. Changed the on hover affect to make it darker, like it does on the real Google.

Added Search icon to SearchBox. Will float left on hover/focus.

It's not perfect, but it's a lot closer than before. Icon should really be grey but I didn't want to get the photo editor out.